+++ src/ml_window.c
@@ -1927,7 +1927,7 @@ ml_window_get_bg_color(
 int
 ml_window_fade(
 	ml_window_t *  win ,
-	u_int8_t  fade_ratio
+	u_int  fade_ratio	/* valid value is 0 - 99 */
 	)
 {
 	x_color_t *  xcolor ;
